Only reexport headers if version is set
When building different devices (non-SODP) in the same tree the
following error shows up:

    error: kernel/sony/qti-headers/Android.bp:31:1: "qti_kernel_headers" depends on undefined module "qti_kernel_headers_"

As per the documentaition `.String()` on `VendorConfig` returns an empty
string if the variable was not set, leading to the `qti_kernel_headers_`
name.  Fix this by only reexporting these headers when a version is set.

Signed-off-by: Marijn Suijten <marijn.suijten@somainline.org>